requirements.txt
cobs
opencv-python
pyserial
scipy
